updated geocoder component (and related stuff)
